tm_parser
Module dot_generation

Module to import when you want to generate DOT code from Turnip code.

This module is made to be interfaced with WebAssembly from a browser.

Functions

dot_from_state 🔒

Private utility function to transform a given state into DOT graph transitions

tm_string_to_dot

Takes a TM machine (.tm) code and turns it into a .dot graph code.