tm_parser
Module parser

Module to import when you want to parse Turing machines code

The defined functions make use of the pest parser generator.

Modules

turnip_v0 🔒

Turnip V0 parser functions

turnip_v1 🔒

Turnip V1 parser functions

turnip_v2 🔒

Turnip V2 parser functions

Structs

ReadEnv

Read symbols struct of a transition object.

State

State struct

Transition

Struct for a transition between 2 states.

WriteFun

Function object applied during a transition.

WritePair

Struct for a Single write action on a TM tape.

Enums

WriteEnv

Struct representing actions performed by a transition.

Functions

get_parsed_file

Get the parsed file We use pest to do the lexing and parsing