tm_parser::turing_machines
Enum Fun

pub enum Fun {

    MvMain(i32),
    MvWork(i32),
    WriteMain(Gamma),
    WriteWork(Gamma),
    IncrBitsMain(),
    IncrBitsWork(),
    DecrBitsMain(),
    DecrBitsWork(),
    BitwiseNotMain(),
    BitwiseNotWork(),
    Add(),
    Sub(),
    Mul(),
    Mod(),
    Div(),
    BitwiseOr(),
    BitwiseAnd(),
    BitwiseXor(),
    BitwiseNand(),
    Geq(),
    Leq(),
    Gt(),
    Lt(),
    Eq(),
    Neq(),

}

Instead of Base Actions, we can use a list of Functions to apply. Fun holds the implementation of all available functions.

Variants

MvMain(i32)

Moves the main tape head by a (neg or pos) amount of cells.

MvWork(i32)

Moves the work tape head by a (neg or pos) amount of cells.

WriteMain(Gamma)

Writes the given letter from Gamma to the main tape at the current head position.

WriteWork(Gamma)

Writes the given letter from Gamma to the work tape at the current head position.

IncrBitsMain()

Adds one to the letter currently under the main tape head

IncrBitsWork()

Adds one to the letter currently under the work tape head

DecrBitsMain()

Subtracts one to the letter currently under the main tape head

DecrBitsWork()

Subtracts one to the letter currently under the work tape head

BitwiseNotMain()

Performs a bitwise not on the letter currently under the main tape head

BitwiseNotWork()

Performs a bitwise not on the letter currently under the work tape head

Add()

Reads the next two letters A and B on the main tape, puts A + B in the third position, and goes back to the head position in which it was before this function

Sub()

Reads the next two letters A and B on the main tape, puts A - B in the third position, and goes back to the head position in which it was before this function

Mul()

Reads the next two letters A and B on the main tape, puts A * B in the third position, and goes back to the head position in which it was before this function

Mod()

Reads the next two letters A and B on the main tape, puts A % B in the third position, and goes back to the head position in which it was before this function

Div()

Reads the next two letters A and B on the main tape, puts A / B in the third position, and goes back to the head position in which it was before this function

BitwiseOr()

Performs a bitwise or on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseAnd()

Performs a bitwise and on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseXor()

Performs a bitwise xor on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseNand()

Performs a bitwise nand on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

Geq()

Compares the next two letters A and B on the main tape, if A >= B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Leq()

Compares the next two letters A and B on the main tape, if A <= B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Gt()

Compares the next two letters A and B on the main tape, if A > B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Lt()

Compares the next two letters A and B on the main tape, if A < B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Eq()

Compares the next two letters A and B on the main tape, if A = B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Neq()

Compares the next two letters A and B on the main tape, if A != B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Implementations

impl Fun

fn eval( &self, pos_main: &mut TapePos, _tape_main: &mut [Gamma], pos_work: &mut TapePos, _tape_work: &mut [Gamma], ) -> Vec<TapeEdit>