tm_parser::turing_machines
Struct Simu

pub struct Simu {
    _tm: TM,
    _cur_state: State,
    _head_pos_main: TapePos,
    _tape_main: Vec<Gamma>,
    _head_pos_work: TapePos,
    _tape_work: Vec<Gamma>,
    _past_edits: Vec<TmEdit>,
    _future_edits: Vec<TmEdit>,
}

Turing Machine simulation object. This object is made to be usable from JavaScript in the web via WebAssembly.

Fields
_tm: TM_cur_state: State_head_pos_main: TapePos_tape_main: Vec<Gamma>_head_pos_work: TapePos_tape_work: Vec<Gamma>_past_edits: Vec<TmEdit>_future_edits: Vec<TmEdit>
Implementations

impl Simu

pub fn new( input_string: &str, grammar_version: i8, main_tape: Vec<Gamma>, work_tape: Vec<Gamma>, fun_env: Vec<String>, ) -> Result<Simu, String>

Simulation object constructor.

After this call, the returned object can directly be used to run the parsed machine.

impl Simu

fn apply_tm_edit(&mut self, edit: &mut TmEdit)

impl Simu

pub fn is_start(&self) -> bool

Checks if the Simulation’s TM has not started yet. NB: this is NOT a check for the START state.

pub fn is_end(&self) -> bool

Checks if the Simulation’s TM has reached the END state.

pub fn is_error(&self) -> bool

Checks if the Simulation’s TM has reached the ERROR state.

impl Simu

pub fn next_step(&mut self)

Runs a single step (i.e. takes a single transition) of the simulated Turing Machine.

pub fn prev_step(&mut self)

Rewinds the Turing Machine one step back.

pub fn all_steps(&mut self)

Runs the whole Turing Machine for a maxiumum number of iterations.

We simply call _next_step in a while not finished loop.

impl Simu

pub fn edit_main_tape()

TODO: documentation

pub fn edit_work_tape()

TODO: documentation

impl Simu

pub fn verify_output(&self, expected: &[Gamma]) -> bool

Verifies that the current main tape has the expected result given in arguments.

Returns true if on END state and the expected output matches the main tape.

impl Simu

pub fn verify_work_tape(&self, expected: &[Gamma]) -> bool

Helper function for tests that enables checking the content of the work tape given an expected content.

impl Simu

pub fn get_current_state(&self) -> String

Returns the current state ID of the simulated Turing Machine.

pub fn head_pos_main(&self) -> TapePos

Returns the current position of the main tape’s head of the simulated TM.

pub fn head_pos_work(&self) -> TapePos

Returns the current position of the work tape’s head of the simulated TM.

pub fn get_main_tape(&self) -> Vec<Gamma>

Returns the main tape

pub fn get_work_tape(&self) -> Vec<Gamma>

Returns the work tape