tm_parser::turing_machines
Enum Fun

pub enum Fun {

    MvMain(i32),
    MvWork(i32),
    WriteMain(Gamma),
    WriteWork(Gamma),
    IncrMainBits(),
    IncrMainByte(),
    IncrWorkBits(),
    IncrWorkByte(),
    DecrMainBits(),
    DecrMainByte(),
    DecrWorkBits(),
    DecrWorkByte(),
    BitwiseNotMainBits(),
    BitwiseNotMainByte(),
    BitwiseNotWorkBits(),
    BitwiseNotWorkByte(),
    AddBits(),
    AddByte(),
    SubBits(),
    SubByte(),
    MulBits(),
    MulByte(),
    ModBits(),
    ModByte(),
    DivBits(),
    DivByte(),
    BitwiseOrBits(),
    BitwiseOrByte(),
    BitwiseAndBits(),
    BitwiseAndByte(),
    BitwiseXorBits(),
    BitwiseXorByte(),
    BitwiseNandBits(),
    BitwiseNandByte(),
    GeqBits(),
    GeqByte(),
    LeqBits(),
    LeqByte(),
    GtBits(),
    GtByte(),
    LtBits(),
    LtByte(),
    EqBits(),
    EqByte(),
    NeqBits(),
    NeqByte(),

}

Instead of Base Actions, we can use a list of Functions to apply. Fun holds the implementation of all available functions.

Variants

MvMain(i32)

Moves the main tape head by a (neg or pos) amount of cells.

MvWork(i32)

Moves the work tape head by a (neg or pos) amount of cells.

WriteMain(Gamma)

Writes the given letter from Gamma to the main tape at the current head position.

WriteWork(Gamma)

Writes the given letter from Gamma to the work tape at the current head position.

IncrMainBits()

TODO

IncrMainByte()

Adds one to the letter currently under the main tape head

IncrWorkBits()

TODO

IncrWorkByte()

Adds one to the letter currently under the work tape head

DecrMainBits()

TODO

DecrMainByte()

Subtracts one to the letter currently under the main tape head

DecrWorkBits()

TODO

DecrWorkByte()

Subtracts one to the letter currently under the work tape head

BitwiseNotMainBits()

TODO

BitwiseNotMainByte()

Performs a bitwise not on the letter currently under the main tape head

BitwiseNotWorkBits()

TODO

BitwiseNotWorkByte()

Performs a bitwise not on the letter currently under the work tape head

AddBits()

TODO

AddByte()

Reads the next two letters A and B on the main tape, puts A + B in the third position, and goes back to the head position in which it was before this function

SubBits()

TODO

SubByte()

Reads the next two letters A and B on the main tape, puts A - B in the third position, and goes back to the head position in which it was before this function

MulBits()

TODO

MulByte()

Reads the next two letters A and B on the main tape, puts A * B in the third position, and goes back to the head position in which it was before this function

ModBits()

TODO

ModByte()

Reads the next two letters A and B on the main tape, puts A % B in the third position, and goes back to the head position in which it was before this function

DivBits()

TODO

DivByte()

Reads the next two letters A and B on the main tape, puts A / B in the third position, and goes back to the head position in which it was before this function

BitwiseOrBits()

TODO

BitwiseOrByte()

Performs a bitwise or on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseAndBits()

TODO

BitwiseAndByte()

Performs a bitwise and on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseXorBits()

TODO

BitwiseXorByte()

Performs a bitwise xor on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

BitwiseNandBits()

TODO

BitwiseNandByte()

Performs a bitwise nand on the next two letters on the main tape, puts the result in the third position, and goes back to the head position in which it was before this function

GeqBits()

TODO

GeqByte()

Compares the next two letters A and B on the main tape, if A >= B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

LeqBits()

TODO

LeqByte()

Compares the next two letters A and B on the main tape, if A <= B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

GtBits()

TODO

GtByte()

Compares the next two letters A and B on the main tape, if A > B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

LtBits()

TODO

LtByte()

Compares the next two letters A and B on the main tape, if A < B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

EqBits()

TODO

EqByte()

Compares the next two letters A and B on the main tape, if A = B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

NeqBits()

TODO

NeqByte()

Compares the next two letters A and B on the main tape, if A != B, puts 1 in the third position, otherwise puts 0. Finally, goes back to the head position in which it was before this function

Implementations

impl Fun

fn eval( &self, pos_main: &mut TapePos, _tape_main: &mut [Gamma], pos_work: &mut TapePos, _tape_work: &mut [Gamma], ) -> Vec<TapeEdit>